requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=10.0.0
numpy>=1.24.0
chromadb>=0.4.0
streamlit>=1.28.0
kagglehub
requests>=2.31.0